--- src/pages/ContactsPage/TeamInfoArray.jsx
@@ -1,59 +1,0 @@
-const TeamInfoArray = [
-  {
-    avatar: 'https://avatars.githubusercontent.com/u/32988243?v=4',
-    name: 'Sergiy Sierov',
-    position: 'Team Lead',
-    id: 1,
-    gitUrl: 'https://github.com/serovsergey',
-    mail: 'serovsergey@gmail.com',
-  },
-  {
-    avatar: 'https://avatars.githubusercontent.com/u/15800167?v=4',
-    name: 'Mariia Chvorun',
-    position: 'Scrum Master',
-    id: 2,
-    gitUrl: 'https://github.com/MariiaC',
-    mail: 'mariiachvorun@gmail.com',
-  },
-  {
-    avatar: 'https://avatars.githubusercontent.com/u/28190277?v=4',
-    name: 'Maksym Bay',
-    position: 'Front-End Developer',
-    id: 3,
-    gitUrl: 'https://github.com/maxxbay',
-    mail: 'maxxbay@gmail.com',
-  },
-  {
-    avatar: 'https://avatars.githubusercontent.com/u/104570430?v=4',
-    name: 'Pavlo Voloshyn',
-    position: 'Front-End Developer',
-    id: 4,
-    gitUrl: 'https://github.com/VPagent',
-    mail: 'vpagent13@gmail.com',
-  },
-  {
-    avatar: 'https://avatars.githubusercontent.com/u/104676845?v=4',
-    name: 'Daryna Shvydiuk',
-    position: 'Front-End Developer',
-    id: 5,
-    gitUrl: 'https://github.com/Darina-Shvydiuk',
-    mail: 'dariyashydiuk@gmail.com',
-  },
-  {
-    avatar: 'https://avatars.githubusercontent.com/u/104834083?v=4',
-    name: 'Roman Serediuk',
-    position: 'Front-End Developer',
-    id: 6,
-    gitUrl: 'https://github.com/isITnow',
-    mail: 'box4rom@gmail.com',
-  },
-  {
-    avatar: 'https://avatars.githubusercontent.com/u/104681622?v=4',
-    name: 'Oleksandr Moshynskiy',
-    position: 'Front-End Developer',
-    id: 7,
-    gitUrl: 'https://github.com/shuriuss',
-    mail: '4030606@gmail.com',
-  },
-];
-export default TeamInfoArray;
